vbd-visionaidashboard-api/requirements/base.txt
django==4.0.1
djangorestframework==3.13.0
bcrypt==3.1.4
django-debug-toolbar==3.2.4
mysqlclient==2.1.0
pillow==9.0.1
gunicorn==20.0.4
django-filter==21.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.11.0
pandas
iteration_utilities
openpyxl
xlrd
django-rest-swagger